Refactor: Split up some modules + rename as .jsx files
