+++ .neutrinorc.js
@@ -17,7 +17,6 @@ module.exports = {
         eslint: {
           rules: {
             "no-console": "off",
-            "react/jsx-filename-extension": [1, { "extensions": [".js", ".jsx"] }],
             "react/prop-types": "off",
           }
         }
