+++ src/containers/fileViewer.jsx
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import * as queryString from 'query-string';
 
 import { fileRevisionCoverageSummary, fileRevisionWithActiveData, rawFile } from '../utils/data';
-import { TestsSideViewer, CoveragePercentageViewer } from './fileviewercov';
+import { TestsSideViewer, CoveragePercentageViewer } from '../components/fileViewer';
 import { HORIZONTAL_ELLIPSIS, HEAVY_CHECKMARK } from '../utils/symbol';
 import hash from '../utils/hash';
